refactor[#175] : 사용자의 현재 위치와 params를 관리하는 컴포넌트 생성
