+++ src/widgets/map/ui/index.ts
@@ -2,3 +2,4 @@ export * from "./GoogleMapsCopyRight";
 export * from "./GoogleMaps";
 export * from "./MapControlWidget";
 export * from "./MapMarkerWidget";
+export * from "./MapInitializer";
